+++ src/modules/p2tg/PostSender.php
@@ -206,9 +206,15 @@ class PostSender extends BaseClass {
 					$this->form_data['delay'] = MainUtils::sanitize( $_POST[ Main::PREFIX . 'delay' ], true ); // phpcs:ignore
 				}
 
-				// if notifications to be disabled.
+				// if notifications are to be disabled.
 				if ( isset( $_POST[ Main::PREFIX . 'disable_notification' ] ) ) { // phpcs:ignore
 					$this->form_data['disable_notification'] = true;
+				}
+
+				// if send featured image.
+				if ( isset( $_POST[ Main::PREFIX . 'send_featured_image' ] ) ) { // phpcs:ignore
+					$send_featured_image = MainUtils::sanitize( $_POST[ Main::PREFIX . 'send_featured_image' ] ); // phpcs:ignore
+					$this->form_data['send_featured_image'] = 'on' === $send_featured_image;
 				}
 			}
 		}
@@ -237,7 +243,7 @@ class PostSender extends BaseClass {
 			'inline_button_url'        => '',
 			'inline_url_button'        => false,
 			'message_template'         => '',
-			'parse_mode'               => null,
+			'parse_mode'               => '',
 			'plugin_posts'             => false,
 			'post_types'               => $array,
 			'rules'                    => $array,
@@ -891,6 +897,11 @@ class PostSender extends BaseClass {
 				$options['delay'] = $this->form_data['delay'];
 			}
 
+			// if send_featured_image overridden.
+			if ( isset( $this->form_data['send_featured_image'] ) ) {
+				$options['send_featured_image'] = $this->form_data['send_featured_image'];
+			}
+
 			// if notifications to be disabled.
 			$options['disable_notification'] = ! empty( $this->form_data['disable_notification'] );
 		}
@@ -1096,7 +1107,7 @@ class PostSender extends BaseClass {
 					preg_match( '/.{1,1024}(?=\s|$)/us', $text, $match );
 					$caption = $match[0];
 
-				} elseif ( 'after' === $image_position && null !== $parse_mode ) {
+				} elseif ( 'after' === $image_position && '' !== $parse_mode ) {
 
 					$text = $this->add_hidden_image_url( $text, $image_source, $parse_mode );