+++ changelog.md
@@ -4,6 +4,10 @@ All notable changes to this project are documented in this file.
 
 ## Unreleased
 
+### Enhancements
+
+- Added override option for Send Featured Image.
+
 ## [3.1.4 - 2021-10-17](https://github.com/manzoorwanijk/wptelegram/releases/tag/v3.1.4)
 
 ### Bug fixes